deleteAlert implemented with prior confirmation to accept

+++ src/components/ConfirmDelete.tsx
@@ -1,17 +1,24 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 "use client";
 import { useState } from "react";
-import { deleteUser, deleteDebt, deletePayment } from "../lib/api";
-import type { Debt, Payment, User } from "../types";
+import { deleteUser, deleteDebt, deletePayment, deleteAlert } from "../lib/api";
+import type { Alert, Debt, Payment, User } from "../types";
 
 interface ConfirmDeleteProps {
     setVisibleConfirmDelete?: (visible: boolean) => void;
     userDelete?: User;
     debtDelete?: Debt;
     paymentDelete?: Payment;
+    alertDelete?: Alert;
 }
 
-export default function ConfirmDelete({ setVisibleConfirmDelete, userDelete, debtDelete, paymentDelete }: ConfirmDeleteProps) {
+export default function ConfirmDelete({ 
+    setVisibleConfirmDelete, 
+    userDelete, 
+    debtDelete, 
+    paymentDelete,
+    alertDelete
+}: ConfirmDeleteProps) {
     
     
     const [loading, setLoading] = useState(false);
@@ -23,14 +30,17 @@ export default function ConfirmDelete({ setVisibleConfirmDelete, userDelete, deb
         setMessage(null);
 
         try {
-            if(userDelete && !debtDelete && !paymentDelete) {
+            if(userDelete && !debtDelete && !paymentDelete && !alertDelete) {
                 await deleteUser(userDelete._id);
             } 
-            if(!userDelete && debtDelete && !paymentDelete) {
+            if(!userDelete && debtDelete && !paymentDelete && !alertDelete) {
                 await deleteDebt(debtDelete._id);
             } 
-            if(!userDelete && !debtDelete && paymentDelete) {
+            if(!userDelete && !debtDelete && paymentDelete && !alertDelete) {
                 await deletePayment(paymentDelete._id);
+            } 
+            if(!userDelete && !debtDelete && !paymentDelete && alertDelete) {
+                await deleteAlert(alertDelete._id);
             } 
             // console.log("Created user:", response);
 
@@ -50,7 +60,8 @@ export default function ConfirmDelete({ setVisibleConfirmDelete, userDelete, deb
                 Delete 
                 {userDelete && " User"} 
                 {debtDelete && " Debt"}
-                {paymentDelete && " Payment"} 
+                {paymentDelete && " Payment"}
+                {alertDelete && " Alert"} 
             </h2>
             
             <form onSubmit={handleSubmit} className={`
@@ -59,8 +70,9 @@ export default function ConfirmDelete({ setVisibleConfirmDelete, userDelete, deb
                 ${userDelete && "bg-blue-300"} 
                 ${debtDelete && "bg-green-300"} 
                 ${paymentDelete && "bg-yellow-200"} 
+                ${alertDelete && "bg-red-200"} 
                 text-gray-800`}>
-                {(userDelete?._id && !debtDelete?._id && !paymentDelete?._id) && ( 
+                {(userDelete?._id && !debtDelete?._id && !paymentDelete?._id && !alertDelete?._id) && ( 
                     <>
                         <label className="input-label">Email: {userDelete?.email}</label>
                         <label className="input-label">Phone: {userDelete?.phone}</label>
@@ -69,7 +81,7 @@ export default function ConfirmDelete({ setVisibleConfirmDelete, userDelete, deb
                         <label className="input-label">Deleted: {userDelete?.deleted}</label>
                     </>
                 ) }
-                {(!userDelete?._id && debtDelete?._id && !paymentDelete?._id) && ( 
+                {(!userDelete?._id && debtDelete?._id && !paymentDelete?._id && !alertDelete?._id) && ( 
                     <>
                         <label className="input-label">
                             ID Creditor: {typeof debtDelete?.id_user_creditor === "object"
@@ -92,14 +104,23 @@ export default function ConfirmDelete({ setVisibleConfirmDelete, userDelete, deb
                         <label className="input-label">Deleted: {debtDelete?.deleted}</label>
                     </>
                 ) }
-                {(!userDelete?._id && !debtDelete?._id && paymentDelete?._id) && ( 
+                {(!userDelete?._id && !debtDelete?._id && paymentDelete?._id && !alertDelete?._id) && ( 
                     <>
                         <label className="input-label">ID Debt: {paymentDelete?.id_debt}</label>
                         <label className="input-label">Amount: {paymentDelete?.amount}</label>
                         <label className="input-label">Date Payment: {paymentDelete?.date_payment}</label>
                         <label className="input-label">Dolar Google: {paymentDelete?.dolar_google}</label>
-                        <label className="input-label">enabled: {paymentDelete?.enabled}</label>
+                        <label className="input-label">Enabled: {paymentDelete?.enabled}</label>
                         <label className="input-label">Deleted: {paymentDelete?.deleted}</label>
+                    </>
+                ) }
+                {(!userDelete?._id && !debtDelete?._id && !paymentDelete?._id && alertDelete?._id) && ( 
+                    <>
+                        <label className="input-label">ID Debt: {alertDelete?.id_debt}</label>
+                        <label className="input-label">Date Alert: {alertDelete?.date_alert}</label>
+                        <label className="input-label">Sent: {alertDelete?.sent}</label>
+                        <label className="input-label">Enabled: {alertDelete?.enabled}</label>
+                        <label className="input-label">Deleted: {alertDelete?.deleted}</label>
                     </>
                 ) }
